Save a layer to a specific path if exist and manage errors message

+++ public/php/data.php
@@ -3,10 +3,14 @@
 $fichier=$_POST['filename'];
 //get json content
 $geojsonStr=$_POST['content'];
-//set output folder
-$destFolder="../data/layers";
+// get folder to save file
+$destFolder=$_POST['path'];
 //save to disk
-file_put_contents($destFolder . DIRECTORY_SEPARATOR .$fichier, $geojsonStr);
-header('Content-type: application/json',true);
-echo '{"success":true, "filepath":"'.$destFolder. DIRECTORY_SEPARATOR .$fichier.'", "geojsonString":'.$geojsonStr.'}';
+if(isset($destFolder) && file_exists($destFolder)) {;
+    file_put_contents($destFolder . DIRECTORY_SEPARATOR .$fichier, $geojsonStr);
+    header('Content-type: application/json',true);
+    echo '{"success":true, "filepath":"'.$destFolder. DIRECTORY_SEPARATOR .$fichier.'", "geojsonString":'.$geojsonStr.'}';
+} else {
+    echo '{"success":false, "error":"Fail to save data to path '.$destFolder.'"}';
+}
 ?>
